Create scraper.py
This is a super-simplified program that retrieves 6 subject grades and then graphs the distribution.  The code is self-explanatory and may not be super-efficient. 

NOTE: Please do not use this program to send multiple requests to the VTU server in order to DoS it. I only created this program for educational purposes .